+++ NEWS.md
@@ -1,8 +1,12 @@
 # TTV
 
+## TTV 0.0.0.6 
+* 29jan2026
+  + added function create_style_table
+
 ## TTV 0.0.0.5 
 * 26jan2026
-  + added argumentshtml_inclde and caption to create_html_table
+  + added arguments html_include and caption to create_html_table
 
 ## TTV 0.0.0.4 
 * 14aug2023